refactor(#109): contractchange 경로 변경 되돌리기
dev 브랜치가 contractchange 패키지를 계속 사용 중이라 merge 시
경로 충돌이 발생하여, 본인인증/stepper 통일 등 기능 변경은 유지한 채
domain.contract로 옮겼던 contractchange 관련 클래스/리소스만
domain.contractchange로 원복
