+++ src/main/java/org/teamsai/saibackend/domain/contractchange/dto/ChangeRequestStatus.java
@@ -1,4 +1,4 @@
-package org.teamsai.saibackend.domain.contract.dto;
+package org.teamsai.saibackend.domain.contractchange.dto;
 
 public enum ChangeRequestStatus {
     PENDING,
