+++ src/main/resources/static/js/contractchange/contract-edit-complete.js
