+++ src/main/resources/static/js/contractchange/contractchange.js
