+++ src/main/resources/templates/contractchange/contract-edit-complete.html
@@ -64,6 +64,6 @@
     </footer>
 </main>
 
-<script th:src="@{/js/contract/contract-edit-complete.js}" src="../../static/js/contract/contract-edit-complete.js"></script>
+<script th:src="@{/js/contractchange/contract-edit-complete.js}" src="../../static/js/contractchange/contract-edit-complete.js"></script>
 </body>
 </html>
