+++ src/main/resources/templates/contractchange/request.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>계약 조건 변경 요청</title>
-    <link rel="stylesheet" th:href="@{/css/contract/contractchange.css}">
+    <link rel="stylesheet" th:href="@{/css/contractchange/contractchange.css}">
 </head>
 <body>
 
@@ -54,6 +54,6 @@
     </div>
 </div>
 
-<script th:src="@{/js/contract/contractchange.js}"></script>
+<script th:src="@{/js/contractchange/contractchange.js}"></script>
 </body>
 </html>
